requirements.txt
aiosqlite==0.19.0
annotated-types==0.6.0
anyio==3.7.1
certifi==2023.11.17
click==8.1.7
exceptiongroup==1.2.0
fake-useragent==1.4.0
fastapi==0.104.1
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
loguru==0.7.2
numpy==1.26.2
orjson==3.9.10
pandas==2.1.3
psycopg==3.1.13
psycopg-pool==3.2.0
pydantic==2.5.2
pydantic_core==2.14.5
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
schedule==1.2.1
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
twscrape==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
uvicorn==0.24.0.post1
